Improve CPU offload

+++ auto_gptq/nn_modules/qlinear.py
@@ -193,7 +193,6 @@ class QuantLinear(nn.Module):
                 autogptq_cuda.vecquant8matmul(x.float(), self.qweight, out, self.scales.float(), self.qzeros, self.g_idx)
             else:
                 raise NotImplementedError("Only 2,3,4,8 bits are supported.")
-            out = out.half()
         else:
             if self.wf.device != self.qzeros.device:
                 self.wf = self.wf.to(self.qzeros.device)
@@ -252,7 +251,7 @@ class QuantLinear(nn.Module):
                     weights.append(scale_i[g_idx_i.long()] * (weight_i - zeros_i[g_idx_i.long()]))
                 weights = torch.cat(weights,dim=1)
             out = torch.matmul(x.half(), weights)
-        out = out.reshape(out_shape)
+        out = out.half().reshape(out_shape)
         out = out + self.bias if self.bias is not None else out
         return out
